Add feedback, circuitpython download link, read-only fixes

+++ run.py
@@ -1,5 +1,4 @@
 import ctypes
-import hashlib
 import json
 import os
 import re
@@ -8,7 +7,6 @@ import shutil
 import string
 import sys
 import tempfile
-from urllib.request import urlretrieve
 import zipfile
 
 def get_readme_content(version):
@@ -33,7 +31,6 @@ def read_board_info(file_path: str) -> str:
         return file.read()
 
 def extract_board_id(board_info: str):
-    # Extract board ID using regular expression
     board_id_search = re.search(r'Board ID:(.+)', board_info)
     
     if board_id_search:
@@ -57,13 +54,9 @@ def get_circuitpy_drive():
                     return f"{letter}:\\"
     return None
 
-def is_drive_writable(drive_path):
-    return os.access(drive_path, os.W_OK)
-
 def get_valid_releases():
     api_url = "https://api.github.com/repos/mechawrench/wificom-lib/releases"
     releases = requests.get(api_url).json()
-
     latest_release = None
     latest_pre_release = None
 
@@ -92,18 +85,34 @@ def is_drive_writable(drive_path):
     else:
         return os.access(drive_path, os.W_OK)
 
+def extract_all_from_archive(archive_path, extract_path):
+    with zipfile.ZipFile(archive_path, 'r') as zip_ref:
+        total_files = len(zip_ref.infolist())
+        extracted_files = 0
+        for file_info in zip_ref.infolist():
+            if not any(part.startswith(".") for part in file_info.filename.split("/")):
+                zip_ref.extract(file_info, extract_path)
+                extracted_files += 1
+                print(f"Extracting: {extracted_files}/{total_files} files", end='\r')
+
+    print("\nExtraction completed.")
+
 def delete_folders_in_lib(destination_folder, lib_folders_to_delete):
     lib_path = os.path.join(destination_folder, "lib")
     for folder in lib_folders_to_delete:
         folder_path = os.path.join(lib_path, folder)
-        if os.path.exists(folder_path) and os.path.isdir(folder_path):
-            shutil.rmtree(folder_path)
+        try:
+            if os.path.exists(folder_path) and os.path.isdir(folder_path) and not any(part.startswith(".") for part in folder_path.split("/")):
+                shutil.rmtree(folder_path)
+        except Exception as e:
+            print(f"An error occurred while deleting {folder_path}: {e}")
+
 
 def get_folders_in_lib(zip_file_path):
     with zipfile.ZipFile(zip_file_path, 'r') as zip_ref:
         lib_folders = set()
         for file_info in zip_ref.infolist():
-            if file_info.filename.startswith("lib/") and file_info.is_dir():
+            if file_info.filename.startswith("lib/") and file_info.is_dir() and not any(part.startswith(".") for part in file_info.filename.split("/")):
                 folder_name = file_info.filename.split('/')[1]
                 lib_folders.add(folder_name)
         return lib_folders
@@ -111,10 +120,17 @@ def get_folders_in_lib(zip_file_path):
 def download_archive(url, save_path):
     response = requests.get(url, stream=True)
     response.raise_for_status()
+    total_bytes = int(response.headers.get('content-length', 0))
+
+    bytes_downloaded = 0
 
     with open(save_path, "wb") as file:
         for chunk in response.iter_content(chunk_size=8192):
             file.write(chunk)
+            bytes_downloaded += len(chunk)
+            print_download_progress(bytes_downloaded, total_bytes)
+
+    print("\nDownload completed.")
 
     return save_path
 
@@ -206,6 +222,12 @@ def choose_release(releases):
     selected_index = int(input("Select a release: ")) - 1
     return releases[selected_index]
 
+def print_download_progress(bytes_downloaded, total_bytes):
+    progress = int((bytes_downloaded / total_bytes) * 100)
+    print(f"Downloading: {progress}% ({bytes_downloaded}/{total_bytes} bytes)", end='\r')
+
+
+
 def check_circuitpython_key(sources_json_path, board_id, device_type, circuitpython_version):
     with open(sources_json_path, 'r') as f:
         sources = json.load(f)
@@ -223,6 +245,12 @@ def check_circuitpython_key(sources_json_path, board_id, device_type, circuitpyt
         if recommended_circuitpython_version != circuitpython_version:
             print(f"\nThe recommended CircuitPython version for this release is {recommended_circuitpython_version} while you have {circuitpython_version} installed.")
             print("It is advised to upgrade/downgrade your CircuitPython version.")
+            if(board_id == 'arduino_nano_rp2040_connect'):
+                print("If you are using the Arduino Nano RP2040 Connect, you can download the latest UF2 file from here:")
+                print("https://adafruit-circuit-python.s3.amazonaws.com/bin/arduino_nano_rp2040_connect/en_US/adafruit-circuitpython-arduino_nano_rp2040_connect-en_US-" + recommended_circuitpython_version + ".uf2")
+            elif board_id == 'raspberry_pi_pico_w':
+                print("If you are using the Raspberry Pi Pico W, you can download the latest UF2 file from here:")
+                print("https://adafruit-circuit-python.s3.amazonaws.com/bin/raspberry_pi_pico_w/en_US/adafruit-circuitpython-raspberry_pi_pico_w-en_US-" + recommended_circuitpython_version + ".uf2")
             decision = input("Type 'Yes' to continue anyways or press Enter to exit: ").lower()
             
             if decision != 'yes':
@@ -311,10 +339,6 @@ def get_selected_release_and_url(selected_option, valid_releases, all_releases, 
 
     return selected_release, download_url
 
-def extract_all_from_archive(archive_path, extract_path):
-    with zipfile.ZipFile(archive_path, 'r') as zip_ref:
-        zip_ref.extractall(extract_path)
-
 def choose_specific_release(all_releases):
     print("\nAvailable releases:")
     for index, release in enumerate(all_releases):
@@ -328,35 +352,74 @@ def choose_specific_release(all_releases):
         print("Invalid selection. Exiting.")
         sys.exit()
 
+def is_hidden_file(file_name):
+    return file_name.startswith("._")
+
 def copy_files_to_destination(destination_folder, extract_path):
     src_lib_folder = os.path.join(extract_path, "lib")
     dst_lib_folder = os.path.join(destination_folder, "lib")
 
-    for item_name in os.listdir(src_lib_folder):
-        src_item = os.path.join(src_lib_folder, item_name)
-        dst_item = os.path.join(dst_lib_folder, item_name)
-        if os.path.isdir(src_item):
-            if os.path.exists(dst_item):
-                shutil.rmtree(dst_item)
-            shutil.copytree(src_item, dst_item)
-        elif os.path.isfile(src_item):
-            shutil.copy2(src_item, dst_item)
+    total_files = len(os.listdir(src_lib_folder))
+    copied_files = 0
 
-    for item in os.listdir(extract_path):
-        src_item = os.path.join(extract_path, item)
-        dst_item = os.path.join(destination_folder, item)
-        if os.path.isfile(src_item):
-            if item in ('config.py', 'board_config.py'):
-                if not os.path.exists(dst_item):
-                    shutil.copy2(src_item, dst_item)
-            else:
+    for item in os.listdir(src_lib_folder):
+        src_item = os.path.join(src_lib_folder, item)
+        dst_item = os.path.join(dst_lib_folder, item)
+
+        if is_hidden_file(item):
+            continue
+
+        try:
+            if os.path.isdir(src_item):
+                if os.path.exists(dst_item):
+                    shutil.rmtree(dst_item)
+                shutil.copytree(src_item, dst_item)
+            elif os.path.isfile(src_item):
                 shutil.copy2(src_item, dst_item)
+            copied_files += 1
+            print(f"Copying: {copied_files}/{total_files} files", end='\r')
+        except Exception as e:
+            print(f"An error occurred while copying {src_item} to {dst_item}: {e}")
 
+    print("\nCopying completed.")
 
-def main():
+def is_drive_writable(drive_path):
+    return os.access(drive_path, os.W_OK)
+
+def remove_hidden_files(drive_path):
+    if os.path.exists(drive_path):
+        for root, dirs, files in os.walk(drive_path, topdown=False):
+            for file_name in files:
+                if file_name.startswith("._"):
+                    file_path = os.path.join(root, file_name)
+                    try:
+                        os.remove(file_path)
+                    except Exception as e:
+                        print(f"Error while removing {file_path}: {e}")
+            for dir_name in dirs:
+                if dir_name.startswith("._"):
+                    dir_path = os.path.join(root, dir_name)
+                    try:
+                        os.rmdir(dir_path)
+                    except Exception as e:
+                        print(f"Error while removing {dir_path}: {e}")
+
+if __name__ == "__main__":
     os.system('cls' if os.name == 'nt' else 'clear')
+
+    destination_folder = get_circuitpy_drive()
+    if destination_folder is None:
+        print("CIRCUITPY drive not found. Exiting...")
+        sys.exit()
+
+    if not is_drive_writable(destination_folder):
+        print("CIRCUITPY drive is read-only. Please make sure the drive is writable and try again.")
+        sys.exit()
     
     print_welcome_message()
+
+    remove_hidden_files(destination_folder)
+
     destination_folder = get_circuitpy_drive()
     circuitpython_version = read_circuitpython_version_from_boot_out(destination_folder)
     
@@ -372,18 +435,27 @@ def main():
     selected_release, download_url = get_selected_release_and_url(selected_option, valid_releases, all_releases, device_type)
 
     temp_directory = tempfile.mkdtemp()
-    archive_path = download_archive(download_url, os.path.join(temp_directory, selected_release.get('sha', selected_release.get('name', '')).replace('/', '_')))
     
-    lib_folders_to_delete = get_folders_in_lib(archive_path)
-    delete_folders_in_lib(destination_folder, lib_folders_to_delete)
+    archive_path = download_archive(download_url, os.path.join(temp_directory, selected_release.get('sha', selected_release.get('name', '')).replace('/', '_')))
+
+    print("\nDownloading archive completed.")
+
+    print("\nDeleted unnecessary folders in lib directory.")
 
     extract_path = os.path.join(temp_directory, "extracted")
     extract_all_from_archive(archive_path, extract_path)
 
+    print("\nArchive extraction completed.")
+
     sources_json_path = os.path.join(extract_path, 'sources.json')
     check_circuitpython_key(sources_json_path, board_id, device_type, circuitpython_version)
 
+    lib_folders_to_delete = get_folders_in_lib(archive_path)
+    delete_folders_in_lib(destination_folder, lib_folders_to_delete)
+
     copy_files_to_destination(destination_folder, extract_path)
+
+    print("\nFiles copying completed.")
 
     shutil.rmtree(temp_directory)
 
@@ -392,6 +464,3 @@ def main():
     print_success_message()
 
     input("Press Enter to exit...")
-
-if __name__ == "__main__":
-    main()
